+++ package.js
@@ -1,5 +1,5 @@
 var options = {
-  version: '1.3.3',
+  version: '1.4.0',
   where: 'client',
   packageName: 'huei90:angular-validation'
 };