+++ src/provider.js
@@ -30,6 +30,12 @@
     var expression = {};
 
     /**
+     * default valid method
+     * @type {{}}
+     */
+    var validMethod = null;
+
+    /**
      * default error, success message
      * @type {{}}
      */
@@ -71,6 +77,23 @@
      */
     this.getDefaultMsg = function(msg) {
       return defaultMsg[msg];
+    };
+
+    /**
+     * allow user to set the global valid method
+     * @param v
+     * @returns {*}
+     */
+    this.setValidMethod = function(v) {
+      validMethod = v;
+    };
+
+    /**
+     * Get the valid method     
+     * @returns {*}
+     */
+    this.getValidMethod = function() {
+      return validMethod;
     };
 
     /**
@@ -210,6 +233,12 @@
     this.invalidCallback = null;
 
     /**
+     * Do this function when reset is performed
+     * @param element
+     */
+    this.resetCallback = null;
+
+    /**
      * reset the specific form
      * @param form
      */
@@ -241,6 +270,8 @@
     this.$get = ['$injector', function($injector) {
       setup($injector);
       return {
+        setValidMethod: this.setValidMethod,
+        getValidMethod: this.getValidMethod,
         setErrorHTML: this.setErrorHTML,
         getErrorHTML: this.getErrorHTML,
         setSuccessHTML: this.setSuccessHTML,
@@ -255,6 +286,7 @@
         validate: this.validate,
         validCallback: this.validCallback,
         invalidCallback: this.invalidCallback,
+        resetCallback: this.resetCallback,
         reset: this.reset
       };
     }];