+++ src/validator.directive.js
@@ -7,6 +7,7 @@
     var $validationProvider = $injector.get('$validation');
     var $q = $injector.get('$q');
     var $timeout = $injector.get('$timeout');
+    var $parse = $injector.get('$parse');
 
     /**
      * Do this function if validation valid
@@ -19,9 +20,12 @@
      */
     var validFunc = function(element, validMessage, validation, scope, ctrl, validations, hasValidation) {
       var messageToShow = validMessage || $validationProvider.getDefaultMsg(validation).success;
+      var validCallback = $parse('success');
+      var messageId = attrs.messageId;
+      var validationGroup = attrs.validationGroup;
       var messageElem;
 
-      if (scope.messageId) messageElem = angular.element(document.querySelector('#' + scope.messageId));
+      if (messageId || validationGroup) messageElem = angular.element(document.querySelector('#' + (messageId || validationGroup)));
       else messageElem = element.next();
 
       if (element.attr('no-validation-message')) {
@@ -33,10 +37,7 @@
         messageElem.css('display', 'none');
       }
 
-      // ctrl.$setValidity(ctrl.$name, true);
-
-      // modified by jacky_coder@163.com
-      // for async issues
+      // For async issues
       scope['valid-rule-' + validation] = true;
 
       if (!hasValidation) {
@@ -47,16 +48,16 @@
           validStatuses.push(scope['valid-rule-' + validator]);
         });
 
-        var isEveryFeildOk = validStatuses.every(function(item) {
+        var isEveryFieldOk = validStatuses.every(function(item) {
           return !!item;
         });
 
-        if (isEveryFeildOk) {
+        if (isEveryFieldOk) {
           ctrl.$setValidity(ctrl.$name, true);
         }
       }
 
-      if (scope.validCallback) scope.validCallback({
+      if (validCallback) validCallback({
         message: messageToShow
       });
       if ($validationProvider.validCallback) $validationProvider.validCallback(element);
@@ -74,11 +75,14 @@
      * @param ctrl
      * @returns {}
      */
-    var invalidFunc = function(element, validMessage, validation, scope, ctrl) {
+    var invalidFunc = function(element, validMessage, validation, scope, ctrl, attrs) {
       var messageToShow = validMessage || $validationProvider.getDefaultMsg(validation).error;
+      var invalidCallback = $parse('error');
+      var messageId = attrs.messageId;
+      var validationGroup = attrs.validationGroup;
       var messageElem;
 
-      if (scope.messageId) messageElem = angular.element(document.querySelector('#' + scope.messageId));
+      if (messageId || validationGroup) messageElem = angular.element(document.querySelector('#' + (messageId || validationGroup)));
       else messageElem = element.next();
 
       if (element.attr('no-validation-message')) {
@@ -90,12 +94,11 @@
         messageElem.css('display', 'none');
       }
 
-      // added by jacky_coder@163.com
-      // for muti conditions
+      // For muti conditions
       scope['valid-rule-' + validation] = false;
 
       ctrl.$setValidity(ctrl.$name, false);
-      if (scope.invalidCallback) scope.invalidCallback({
+      if (invalidCallback) invalidCallback({
         message: messageToShow
       });
       if ($validationProvider.invalidCallback) $validationProvider.invalidCallback(element);
@@ -103,6 +106,33 @@
       return false;
     };
 
+    /**
+     * Verify whether there is one of the elements inside the group valid.
+     * If so, it returns true, otherwise, it returns false
+     *
+     * @param scope
+     * @param element
+     * @param attrs
+     * @param ctrl
+     * @return {boolean}
+     */
+    var checkValidationGroup = function(scope, element, attrs, ctrl) {
+      var validationGroup = attrs.validationGroup;
+      var validationGroupElems = document.querySelectorAll('*[validation-group=' + validationGroup + ']');
+      var validationGroupElem;
+
+      // Set the element to be invalid
+      ctrl.$setValidity(ctrl.$name, false);
+
+      // Loop through all elements inside the group
+      for (var i = 0, len = validationGroupElems.length; i < len; i++) {
+        validationGroupElem = angular.element(validationGroupElems[i]);
+
+        // If the element is valid and it's not the same element with the current checking element, returns true
+        if (validationGroupElem.hasClass('ng-valid') && validationGroupElem[0] !== element[0]) return true;
+      }
+      return false;
+    };
 
     /**
      * collect elements for focus
@@ -128,8 +158,7 @@
       var paramIndex = validatorExpr.indexOf('=');
       var validator = paramIndex === -1 ? validatorExpr : validatorExpr.substr(0, paramIndex);
 
-      // added by jacky_coder@163.com
-      // case: feild when filled must validate or do nothing.
+      // Case: Field when filled must validate, or do nothing.
       var isUnRequired = validator.indexOf('!') > -1 ? true : false;
 
       if (isUnRequired) {
@@ -141,6 +170,7 @@
       var successMessage = validator + 'SuccessMessage';
       var errorMessage = validator + 'ErrorMessage';
       var expression = $validationProvider.getExpression(validator);
+      var validationGroup = attrs.validationGroup;
       var valid = {
         success: function() {
           validFunc(element, attrs[successMessage], validator, scope, ctrl, validators, leftValidation.length);
@@ -151,7 +181,7 @@
           }
         },
         error: function() {
-          return invalidFunc(element, attrs[errorMessage], validator, scope, ctrl);
+          return invalidFunc(element, attrs[errorMessage], validator, scope, ctrl, attrs);
         }
       };
 
@@ -161,13 +191,12 @@
         else return;
       }
 
-      // added by jacky_coder@163.com
       // if user not input return true, else validate
       if (isUnRequired && !value) {
         return valid.success();
       }
 
-      // for form edit, compared with old value
+      // When form edit, compared with old value
       // Case 1: used controllerAs
       // Case 2: used $scope
 
@@ -196,7 +225,11 @@
         return $q.all([$validationProvider.getExpression(validator)(value, scope, element, attrs, validatorParam)])
           .then(function(data) {
             if (data && data.length > 0 && data[0]) return valid.success();
-            else return valid.error();
+            else if (validationGroup) {
+              // Whenever the element is invalid, we'll check whether one of the elements inside the its group valid or not.
+              // If there is a valid element, its invalid message won't be shown, Otherwise, shows its invalid message.
+              if (!checkValidationGroup(scope, element, attrs, ctrl)) valid.error();
+            } else return valid.error();
           }, function() {
             return valid.error();
           });
@@ -206,10 +239,13 @@
       else if (expression.constructor === RegExp) {
         // Only apply the test if the value is neither undefined or null
         if (value !== undefined && value !== null) return $validationProvider.getExpression(validator).test(value) ? valid.success() : valid.error();
-        else return valid.error();
+        else if (validationGroup) {
+          // Whenever the element is invalid, we'll check whether one of the elements inside the its group valid or not.
+          // If there is a valid element, its invalid message won't be shown, Otherwise, shows its invalid message.
+          if (!checkValidationGroup(scope, element, attrs, ctrl)) valid.error();
+        } else return valid.error();
       } else return valid.error();
     };
-
 
     /**
      * generate unique guid
@@ -231,14 +267,19 @@
         validCallback: '&',
         invalidCallback: '&',
         messageId: '@',
-        // added by jacky_coder@163.com
-        // for form edit
-        // readonly field doesn't validate
         editReadonly: '=editReadonly',
-        // when new value equals old, field doesn't validate
         originalData: '=originalData'
       },
       link: function(scope, element, attrs, ctrl) {
+        /**
+         * All attributes
+         */
+        var validator = attrs.validator;
+        var messageId = attrs.messageId;
+        var validationGroup = attrs.validationGroup;
+        var validMethod = attrs.validMethod;
+        var ngModel = attrs.ngModel;
+
         /**
          * watch
          * @type {watch}
@@ -255,7 +296,7 @@
          *
          * Convert user input String to Array
          */
-        var validation = attrs.validator.split(',');
+        var validation = validator.split(',');
 
         /**
          * guid use
@@ -274,7 +315,7 @@
         /**
          * Default Valid/Invalid Message
          */
-        if (!scope.messageId) element.after('<span></span>');
+        if (!(messageId || validationGroup)) element.after('<span></span>');
 
         /**
          * Set custom initial validity
@@ -282,8 +323,9 @@
          */
         ctrl.$setValidity(ctrl.$name, initialValidity);
 
-        // added by jacky_coder@163.com
-        // add edit readonly doesn't validate
+        /**
+         * Readonly doesn't validate
+         */
         if (scope.editReadonly) {
           element.attr('readonly', true);
           ctrl.$setValidity(ctrl.$name, true);
@@ -307,14 +349,17 @@
             ctrl.$setPristine();
             ctrl.$setValidity(ctrl.$name, undefined);
             ctrl.$render();
-            if (scope.messageId) angular.element(document.querySelector('#' + scope.messageId)).html('');
+            if (messageId || validationGroup) angular.element(document.querySelector('#' + (messageId || validationGroup))).html('');
             else element.next().html('');
+
+            if ($validationProvider.resetCallback) $validationProvider.resetCallback(element);
           });
         });
 
         /**
          * Check validator
          */
+        validMethod = (angular.isUndefined(validMethod)) ? $validationProvider.getValidMethod() : validMethod;
 
         /**
          * Click submit form, check the validity when submit
@@ -325,12 +370,12 @@
 
           isValid = checkValidation(scope, element, attrs, ctrl, validation, value);
 
-          if (attrs.validMethod === 'submit') {
+          if (validMethod === 'submit') {
             // clear previous scope.$watch
             watch();
-            watch = scope.$watch('model', function(value, oldValue) {
-              value = ctrl.$viewValue;
-
+            watch = scope.$watch(function() {
+              return scope.$eval(ngModel);
+            }, function(value, oldValue) {
               // don't watch when init
               if (value === oldValue) {
                 return;
@@ -366,9 +411,9 @@
         /**
          * Validate blur method
          */
-        if (attrs.validMethod === 'blur') {
+        if (validMethod === 'blur') {
           element.bind('blur', function() {
-            var value = ctrl.$viewValue;
+            var value = scope.$eval(ngModel);
             scope.$apply(function() {
               checkValidation(scope, element, attrs, ctrl, validation, value);
             });
@@ -380,7 +425,7 @@
         /**
          * Validate submit & submit-only method
          */
-        if (attrs.validMethod === 'submit' || attrs.validMethod === 'submit-only') {
+        if (validMethod === 'submit' || validMethod === 'submit-only') {
           return;
         }
 
@@ -388,8 +433,9 @@
          * Validate watch method
          * This is the default method
          */
-        scope.$watch('model', function(value) {
-          value = ctrl.$viewValue;
+        scope.$watch(function() {
+          return scope.$eval(ngModel);
+        }, function(value) {
           /**
            * dirty, pristine, viewValue control here
            */
@@ -398,7 +444,7 @@
             ctrl.$setViewValue(ctrl.$viewValue);
           } else if (ctrl.$pristine) {
             // Don't validate form when the input is clean(pristine)
-            if (scope.messageId) angular.element(document.querySelector('#' + scope.messageId)).html('');
+            if (messageId || validationGroup) angular.element(document.querySelector('#' + (messageId || validationGroup))).html('');
             else element.next().html('');
             return;
           }
@@ -411,7 +457,7 @@
            */
           attrs.$observe('noValidationMessage', function(value) {
             var el;
-            if (scope.messageId) el = angular.element(document.querySelector('#' + scope.messageId));
+            if (messageId || validationGroup) el = angular.element(document.querySelector('#' + (messageId || validationGroup)));
             else el = element.next();
             if (value === 'true' || value === true) el.css('display', 'none');
             else if (value === 'false' || value === false) el.css('display', 'block');